+++ felt/src/bigint_felt.rs
@@ -2,7 +2,7 @@ use lazy_static::lazy_static;
 use num_bigint::{BigInt, BigUint, ToBigInt, U64Digits};
 use num_integer::Integer;
 use num_traits::{Bounded, FromPrimitive, Num, One, Pow, Signed, ToPrimitive, Zero};
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 use std::{
     convert::Into,
     fmt,
@@ -24,7 +24,7 @@ lazy_static! {
         .expect("Conversion BigUint -> BigInt can't fail");
 }
 
-#[derive(Eq, Hash, PartialEq, PartialOrd, Ord, Clone, Deserialize, Default)]
+#[derive(Eq, Hash, PartialEq, PartialOrd, Ord, Clone, Deserialize, Default, Serialize)]
 pub struct FeltBigInt(BigUint);
 
 macro_rules! from_integer {
@@ -307,14 +307,22 @@ impl Sum for FeltBigInt {
 impl Neg for FeltBigInt {
     type Output = FeltBigInt;
     fn neg(self) -> Self::Output {
-        FeltBigInt(&*CAIRO_PRIME - self.0)
+        if self.is_zero() {
+            self
+        } else {
+            FeltBigInt(&*CAIRO_PRIME - self.0)
+        }
     }
 }
 
 impl<'a> Neg for &'a FeltBigInt {
     type Output = FeltBigInt;
     fn neg(self) -> Self::Output {
-        FeltBigInt(&*CAIRO_PRIME - &self.0)
+        if self.is_zero() {
+            self.clone()
+        } else {
+            FeltBigInt(&*CAIRO_PRIME - &self.0)
+        }
     }
 }
 
@@ -354,7 +362,7 @@ impl<'a> Sub for &'a FeltBigInt {
 impl Sub<u32> for FeltBigInt {
     type Output = FeltBigInt;
     fn sub(self, rhs: u32) -> Self {
-        match (&self.0).to_u32() {
+        match (self.0).to_u32() {
             Some(num) if num < rhs => Self(&*CAIRO_PRIME - (rhs - self.0)),
             _ => Self(self.0 - rhs),
         }
@@ -364,7 +372,7 @@ impl Sub<u32> for FeltBigInt {
 impl<'a> Sub<u32> for &'a FeltBigInt {
     type Output = FeltBigInt;
     fn sub(self, rhs: u32) -> Self::Output {
-        match (&self.0).to_u32() {
+        match (self.0).to_u32() {
             Some(num) if num < rhs => FeltBigInt(&*CAIRO_PRIME - (rhs - &self.0)),
             _ => FeltBigInt(&self.0 - rhs),
         }
@@ -374,7 +382,7 @@ impl<'a> Sub<u32> for &'a FeltBigInt {
 impl Sub<usize> for FeltBigInt {
     type Output = FeltBigInt;
     fn sub(self, rhs: usize) -> Self {
-        match (&self.0).to_usize() {
+        match (self.0).to_usize() {
             Some(num) if num < rhs => FeltBigInt(&*CAIRO_PRIME - (rhs - num)),
             _ => FeltBigInt(self.0 - rhs),
         }
@@ -403,7 +411,7 @@ impl Sub<FeltBigInt> for usize {
 impl Sub<&FeltBigInt> for usize {
     type Output = FeltBigInt;
     fn sub(self, rhs: &FeltBigInt) -> Self::Output {
-        match (&rhs.0).to_usize() {
+        match (rhs.0).to_usize() {
             Some(num) => {
                 if num > self {
                     FeltBigInt(&*CAIRO_PRIME - (num - self))
@@ -480,6 +488,7 @@ impl Pow<u32> for FeltBigInt {
 
 impl<'a> Pow<u32> for &'a FeltBigInt {
     type Output = FeltBigInt;
+    #[allow(clippy::needless_borrow)] // the borrow of self.0 is necessary becase it's of the type BigUInt, which doesn't implement the Copy trait
     fn pow(self, rhs: u32) -> Self::Output {
         FeltBigInt((&self.0).pow(rhs).mod_floor(&CAIRO_PRIME))
     }
@@ -492,7 +501,7 @@ impl Div for FeltBigInt {
     fn div(self, rhs: Self) -> Self::Output {
         let x = rhs
             .0
-            .to_bigint() // Always succeeds for BitUint -> BigInt
+            .to_bigint() // Always succeeds for BigUint -> BigInt
             .unwrap()
             .extended_gcd(&CAIRO_SIGNED_PRIME)
             .x;