+++ src/hint_processor/builtin_hint_processor/hint_code.rs
@@ -548,3 +548,7 @@ pub(crate) const RELOCATE_SEGMENT: &str =
     r#"memory.add_relocation_rule(src_ptr=ids.src_ptr, dest_ptr=ids.dest_ptr)"#;
 
 pub(crate) const TEMPORARY_ARRAY: &str = r#"ids.temporary_array = segments.add_temp_segment()"#;
+pub(crate) const VERIFY_ECDSA_SIGNATURE: &str =
+    r#"ecdsa_builtin.add_signature(ids.ecdsa_ptr.address_, (ids.signature_r, ids.signature_s))"#;
+#[cfg(feature = "skip_next_instruction_hint")]
+pub(crate) const SKIP_NEXT_INSTRUCTION: &str = "skip_next_instruction()";